# readium-api/utils/swagger/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Readium API",
    "description": "Handsome backend devs"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
...
  "paths": {
...
    "/dev\r/post": {
      "post": {
        "tags": [
          "Dev"
        ],
        "summary": "Create a post (DEV)",
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "coverImage": {
                    "type": "file"
                  },
                  "title": {
                    "type": "string",
                    "example": "Post Title"
                  },
                  "content": {
                    "type": "string",
                    "example": "Text only content"
                  },
                  "textEditorContent": {
                    "type": "string",
                    "example": "Text Editor Content"
                  },
                  "description": {
                    "type": "string",
                    "example": "Description"
                  },
                  "isPublished": {
                    "type": "string",
                    "example": "false"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users/": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete account",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/users/collections/": {
      "get": {
        "tags": [
          "Collection"
        ],
        "summary": "get all collections of me",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Collection"
        ],
        "summary": "Create new collection",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string",
                    "default": "collection name"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/posts/": {
      "get": {
        "tags": [
          "Post"
        ],
        "summary": "Get posts",
        "description": "",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "type": "string"
          },
          {
            "name": "skip",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}